Rename `amount_model` to `amount`
